Increased panel border sligthly, added production archive.
Increased the PCB border slightly (by 0.2mm on all sides) as some traces
were too close to the edge. This was easier than move the traces as
some are between the boundary and a component (typically an led).

Added production archive for previous revisions of the design.  Included
JLCPCB PCB order details and PCBA info for helping with future orders.